change code -> CreateWallet does not work
